requirements.txt
pysocialforce
matplotlib
paho-mqtt==1.6.1
numpy
pandas
numba==0.58.1
opencv-python
tqdm